requirements.txt
streamlit==1.29.0
pandas==2.2.3
numpy==1.26.4
tensorflow==2.16.1
scikit-learn==1.4.2
gensim==4.3.3
joblib==1.4.0
altair==5.3.0
tinydb==4.8.2
nltk==3.9.1
bcrypt==4.3.0
pyyaml==6.0.1
